[fix] refactor entry package
