+++ app/src/main/java/org/journey/android/entry/OnboardingViewModel.kt
@@ -8,24 +8,5 @@ import javax.inject.Inject
 
 @HiltViewModel
 class OnboardingViewModel @Inject constructor() :BaseViewModel() {
-    private val _accessToken = MutableLiveData<String>()
-    private val _kakaoLogin = MutableLiveData<Boolean>()
-    private val _googleLogin = MutableLiveData<Boolean>()
 
-    val kakaoLogin: LiveData<Boolean>
-        get() = _kakaoLogin
-
-    val googleLogin : LiveData<Boolean>
-        get() = _googleLogin
-
-    val accessToken: LiveData<String>
-        get() = _accessToken
-
-    private val _refreshToken = MutableLiveData<String>()
-    val refreshToken: LiveData<String>
-        get() = _refreshToken
-
-    private val _loginSuccess = MutableLiveData<String>()
-    val loginSuccess: LiveData<String>
-        get() = _loginSuccess
 }
