+++ app/src/main/java/org/journey/android/entry/frame/EntryActivity.kt
@@ -6,7 +6,7 @@ import androidx.navigation.NavController
 import androidx.navigation.fragment.NavHostFragment
 import dagger.hilt.android.AndroidEntryPoint
 import org.journey.android.R
-import org.journey.android.databinding.ActivityOnboardingBinding
+import org.journey.android.databinding.ActivityEntryBinding
 import org.journey.android.preference.SharedPreferencesManager
 import org.journey.android.util.TedRxKeyboardObserver
 import javax.inject.Inject
@@ -15,11 +15,11 @@ import javax.inject.Inject
 class EntryActivity : AppCompatActivity() {
     @Inject
     lateinit var pref: SharedPreferencesManager
-    private lateinit var binding: ActivityOnboardingBinding
+    private lateinit var binding: ActivityEntryBinding
     private lateinit var navController: NavController
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = ActivityOnboardingBinding.inflate(layoutInflater)
+        binding = ActivityEntryBinding.inflate(layoutInflater)
         setContentView(binding.root)
         initNavController()
         setKeyboardObserver()
