+++ app/src/main/res/layout/activity_entry.xml
